Split build_sqlite.py into schema/loaders/orchestrator modules (#1228)
- Create build_sqlite_schema.py (639 lines) with all DDL in create_schema()
- Create build_sqlite_loaders.py (1731 lines) with all populate_* functions
- Slim build_sqlite.py to 196-line orchestrator importing both modules
- python3 _tools/build_sqlite.py produces identical scripture.db
